Fixed bug involved in simultaneous users making transaction.
 Imagine we reserve one copy for each game as a display copy. That is, a non-salable copy. Now let's cook up a scenario where there are only two copies of Diablo II left but our database indicates there is one left because we reserve one for display. Now let's say a customer A grabs the last salable copy and brings it up to a counter, and then right after that another customer, customer B, grabs the display copy and brings it up to a different counter.

Now Employee X is serving customer A, and adds one Diablo II to the cart. The query looks to see if there is at least one copy of Diablo II for sale, sees that there is, and all is good. But, before Employee X does a checkout, Employee Y, who is serving customer B, adds one Diablo II to the shopping cart. The query sees that there is one copy of Diablo II left and all is good so the add is successful.

Now say Employee X does a checkout. After the button is hit pos.games is updated to set Diablo II to quantity 0. Then, unaware that Employee X just sold the last salable copy, Employee Y sales the display copy. Oh no!

 Now when employee hits "Checkout" the database is double checked to make sure the last copy of a game wasn't sold in the time between add item and checkout.